+++ packages/autorest.python/test/vanilla/legacy/Expected/AcceptanceTests/BodyByte/bodybyte/_auto_rest_swagger_bat_byte_service.py
@@ -12,7 +12,7 @@ from typing import Any
 from azure.core import PipelineClient
 from azure.core.rest import HttpRequest, HttpResponse
 
-from . import models
+from . import models as _models
 from ._configuration import AutoRestSwaggerBATByteServiceConfiguration
 from ._serialization import Deserializer, Serializer
 from .operations import ByteOperations
@@ -33,7 +33,7 @@ class AutoRestSwaggerBATByteService:  # pylint: disable=client-accepts-api-versi
         self._config = AutoRestSwaggerBATByteServiceConfiguration(**kwargs)
         self._client = PipelineClient(base_url=base_url, config=self._config, **kwargs)
 
-        client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
+        client_models = {k: v for k, v in _models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
         self._deserialize = Deserializer(client_models)
         self._serialize.client_side_validation = False