+++ packages/autorest.python/test/vanilla/legacy/Expected/AcceptanceTests/BodyComplex/bodycomplex/_auto_rest_complex_test_service.py
@@ -14,7 +14,7 @@ from msrest import Deserializer, Serializer
 from azure.core import PipelineClient
 from azure.core.rest import HttpRequest, HttpResponse
 
-from . import models
+from . import models as _models
 from ._configuration import AutoRestComplexTestServiceConfiguration
 from .operations import (
     ArrayOperations,
@@ -63,7 +63,7 @@ class AutoRestComplexTestService:  # pylint: disable=client-accepts-api-version-
         self._config = AutoRestComplexTestServiceConfiguration(**kwargs)
         self._client = PipelineClient(base_url=base_url, config=self._config, **kwargs)
 
-        client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
+        client_models = {k: v for k, v in _models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
         self._deserialize = Deserializer(client_models)
         self.basic = BasicOperations(self._client, self._config, self._serialize, self._deserialize)