+++ packages/autorest.python/test/vanilla/legacy/Expected/AcceptanceTests/IncorrectErrorResponse/incorrecterrorresponse/aio/operations/_incorrect_returned_error_model_operations.py
@@ -21,7 +21,6 @@ from azure.core.pipeline.transport import AsyncHttpResponse
 from azure.core.rest import HttpRequest
 from azure.core.tracing.decorator_async import distributed_trace_async
 
-from ... import models as _models
 from ..._vendor import _convert_request
 from ...operations._incorrect_returned_error_model_operations import build_get_incorrect_error_from_server_request
 from .._vendor import IncorrectReturnedErrorModelMixinABC