+++ packages/autorest.python/ChangeLog.md
@@ -1,5 +1,21 @@
 # Release History
 
+### 2022-11-xx - 6.2.6
+
+| Library                                                                 | Min Version |
+| ----------------------------------------------------------------------- | ----------- |
+| `@autorest/core`                                                        | `3.9.2`     |
+| `@autorest/modelerfour`                                                 | `4.24.3`    |
+| `azure-core` dep of generated code                                      | `1.24.0`    |
+| `isodate` dep of generated code                                         | `0.6.1`     |
+| `msrest` dep of generated code (If generating legacy code)              | `0.7.1`     |
+| `azure-mgmt-core` dep of generated code (If generating mgmt plane code) | `1.3.2`     |
+| `typing-extensions` dep of generated code (If generating with constants) | `4.0.1`     |
+
+**Bug Fixes**
+
+- Fix clients with enum params  #1558
+
 ### 2022-11-04 - 6.2.5
 
 | Library                                                                 | Min Version |